# openapi.cloudflare.yaml
openapi: 3.0.3
info:
  title: ExitRoute API
  description: A self-hosted informational API. It never authenticates to, operates, or stores credentials
    for a consumer account.
  version: 0.1.0
paths:
...
  /v1/services:
    get:
      tags:
      - routes
      summary: List Services
      operationId: listServices
      security:
      - ApiKeyAuth: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          title: Cursor
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePage'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/services/{service_slug}/exit-route:
    get:
      tags:
      - routes
      summary: Get Current Exit Route
      operationId: getCurrentExitRoute
      security:
      - ApiKeyAuth: []
      parameters:
      - name: service_slug
        in: path
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Service Slug
      - name: region
        in: query
        required: false
        schema:
          type: string
          pattern: ^[A-Z]{2}$
          default: US
          title: Region
      - name: platform
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Platform'
          default: web
      - name: If-None-Match
        in: header
        required: false
        schema:
          title: If-None-Match
          type: string
          nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExitRouteView'
        '304':
          description: The cached representation is current.
          headers:
            ETag:
              description: Strong representation validator.
              schema:
                type: string
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/services/{service_slug}/exit-route/revisions:
    get:
      tags:
      - routes
      summary: List Revisions
      operationId: listExitRouteRevisions
      security:
      - ApiKeyAuth: []
      parameters:
      - name: service_slug
        in: path
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Service Slug
      - name: region
        in: query
        required: false
        schema:
          type: string
          pattern: ^[A-Z]{2}$
          default: US
          title: Region
      - name: platform
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Platform'
          default: web
      - name: cursor
        in: query
        required: false
        schema:
          title: Cursor
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionPage'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/services/{service_slug}/exit-route/revisions/{revision_number}:
    get:
      tags:
      - routes
      summary: Get Revision
      operationId: getExitRouteRevision
      security:
      - ApiKeyAuth: []
      parameters:
      - name: service_slug
        in: path
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Service Slug
      - name: revision_number
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          title: Revision Number
      - name: region
        in: query
        required: false
        schema:
          type: string
          pattern: ^[A-Z]{2}$
          default: US
          title: Region
      - name: platform
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Platform'
          default: web
      - name: If-None-Match
        in: header
        required: false
        schema:
          title: If-None-Match
          type: string
          nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExitRouteView'
        '304':
          description: The cached representation is current.
          headers:
            ETag:
              description: Strong representation validator.
              schema:
                type: string
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/changes:
    get:
      tags:
      - routes
      summary: List Changes
      operationId: listRouteChanges
      security:
      - ApiKeyAuth: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          title: Cursor
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  schemas:
...
    ChangeEventView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        service:
          $ref: '#/components/schemas/ServiceRef'
        variant:
          $ref: '#/components/schemas/Variant'
        revision:
          type: integer
          title: Revision
        event_type:
          type: string
          enum:
          - published
          - superseded
          - stale
          - withdrawn
          title: Event Type
        summary:
          type: string
          title: Summary
      additionalProperties: false
      type: object
      required:
      - id
      - occurred_at
      - service
      - variant
      - revision
      - event_type
      - summary
      title: ChangeEventView
    ChangePage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ChangeEventView'
          type: array
          title: Data
        next_cursor:
          title: Next Cursor
          type: string
          nullable: true
      additionalProperties: false
      type: object
      required:
      - data
      - next_cursor
      title: ChangePage
    Choice:
      properties:
        id:
          type: string
          maxLength: 80
          minLength: 2
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Id
        label:
          type: string
          maxLength: 160
          minLength: 1
          title: Label
        target_node_id:
          type: string
          maxLength: 80
          minLength: 2
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Target Node Id
        effect:
          $ref: '#/components/schemas/ChoiceEffect'
        effort:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Effort
        prominence:
          $ref: '#/components/schemas/Prominence'
      additionalProperties: false
      type: object
      required:
      - id
      - label
      - target_node_id
      - effect
      - effort
      - prominence
      title: Choice
    ChoiceEffect:
      type: string
      enum:
      - advance
      - loop
      - retain
      - handoff
      title: ChoiceEffect
    Confidence:
      type: string
      enum:
      - high
      - medium
      - low
      title: Confidence
...
    EvidenceSummary:
      properties:
        method:
          type: string
          title: Method
          default: manual_clean_session
          enum:
          - manual_clean_session
        verification_sessions:
          type: integer
          title: Verification Sessions
        change_summary:
          type: string
          title: Change Summary
      additionalProperties: false
      type: object
      required:
      - verification_sessions
      - change_summary
      title: EvidenceSummary
    ExitRouteView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        service:
          $ref: '#/components/schemas/ServiceRef'
        outcome:
          $ref: '#/components/schemas/Outcome'
        variant:
          $ref: '#/components/schemas/Variant'
        revision:
          type: integer
          title: Revision
        publication_state:
          $ref: '#/components/schemas/PublicationState'
        status:
          $ref: '#/components/schemas/TrustState'
        verified_at:
          type: string
          format: date-time
          title: Verified At
        review_due_at:
          type: string
          format: date-time
          title: Review Due At
        entry_url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Entry Url
        graph:
          $ref: '#/components/schemas/RouteGraph'
        best_route:
          items:
            type: string
          type: array
          title: Best Route
        friction:
          $ref: '#/components/schemas/Friction'
        confidence:
          $ref: '#/components/schemas/Confidence'
        evidence_summary:
          $ref: '#/components/schemas/EvidenceSummary'
        fingerprint:
          type: string
          title: Fingerprint
      additionalProperties: false
      type: object
      required:
      - id
      - service
      - outcome
      - variant
      - revision
      - publication_state
      - status
      - verified_at
      - review_due_at
      - entry_url
      - graph
      - best_route
      - friction
      - confidence
      - evidence_summary
      - fingerprint
      title: ExitRouteView
    Friction:
      properties:
        screens:
          type: integer
          minimum: 0.0
          title: Screens
        retention_offers:
          type: integer
          minimum: 0.0
          title: Retention Offers
        loops:
          type: integer
          minimum: 0.0
          title: Loops
        offline_handoff:
          type: boolean
          title: Offline Handoff
        effort_score:
          type: integer
          minimum: 0.0
          title: Effort Score
        algorithm_version:
          type: string
          title: Algorithm Version
          default: friction-v1
          enum:
          - friction-v1
      additionalProperties: false
      type: object
      required:
      - screens
      - retention_offers
      - loops
      - offline_handoff
      - effort_score
      title: Friction
...
    NodeKind:
      type: string
      enum:
      - screen
      - os_settings
      - handoff
      - terminal
      title: NodeKind
...
    Outcome:
      type: string
      enum:
      - cancel_subscription
      title: Outcome
    Platform:
      type: string
      enum:
      - web
      - ios
      - android
      - phone
      - mail
      - in_person
      title: Platform
    Prominence:
      type: string
      enum:
      - primary
      - secondary
      - subdued
      - ambiguous
      title: Prominence
    PublicationState:
      type: string
      enum:
      - draft
      - published
      - superseded
      - withdrawn
      title: PublicationState
...
    RevisionPage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/RevisionSummary'
          type: array
          title: Data
        next_cursor:
          title: Next Cursor
          type: string
          nullable: true
      additionalProperties: false
      type: object
      required:
      - data
      - next_cursor
      title: RevisionPage
...
    RevisionSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        revision:
          type: integer
          title: Revision
        publication_state:
          $ref: '#/components/schemas/PublicationState'
        status:
          $ref: '#/components/schemas/TrustState'
        verified_at:
          title: Verified At
          type: string
          format: date-time
          nullable: true
        review_due_at:
          title: Review Due At
          type: string
          format: date-time
          nullable: true
        fingerprint:
          type: string
          title: Fingerprint
        change_summary:
          type: string
          title: Change Summary
      additionalProperties: false
      type: object
      required:
      - id
      - revision
      - publication_state
      - status
      - verified_at
      - review_due_at
      - fingerprint
      - change_summary
      title: RevisionSummary
...
    RouteGraph:
      properties:
        entry_node_id:
          type: string
          maxLength: 80
          minLength: 2
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Entry Node Id
        nodes:
          items:
            $ref: '#/components/schemas/RouteNode'
          type: array
          maxItems: 100
          minItems: 2
          title: Nodes
      additionalProperties: false
      type: object
      required:
      - entry_node_id
      - nodes
      title: RouteGraph
    RouteNode:
      properties:
        id:
          type: string
          maxLength: 80
          minLength: 2
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Id
        kind:
          $ref: '#/components/schemas/NodeKind'
        state:
          $ref: '#/components/schemas/TerminalState'
          nullable: true
        choices:
          items:
            $ref: '#/components/schemas/Choice'
          type: array
          maxItems: 30
          title: Choices
      additionalProperties: false
      type: object
      required:
      - id
      - kind
      - choices
      title: RouteNode
...
    ServicePage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ServiceView'
          type: array
          title: Data
        next_cursor:
          title: Next Cursor
          type: string
          nullable: true
      additionalProperties: false
      type: object
      required:
      - data
      - next_cursor
      title: ServicePage
    ServiceRef:
      properties:
        slug:
          type: string
          title: Slug
        name:
          type: string
          title: Name
      additionalProperties: false
      type: object
      required:
      - slug
      - name
      title: ServiceRef
    ServiceView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        slug:
          type: string
          title: Slug
        name:
          type: string
          title: Name
        domains:
          items:
            type: string
          type: array
          title: Domains
        active:
          type: boolean
          title: Active
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - slug
      - name
      - domains
      - active
      - created_at
      title: ServiceView
    TerminalState:
      type: string
      enum:
      - cancelled
      - retained
      - unavailable
      title: TerminalState
    TrustState:
      type: string
      enum:
      - provisional
      - verified
      - stale
      title: TrustState
...
    Variant:
      properties:
        region:
          type: string
          pattern: ^[A-Z]{2}$
          title: Region
          default: US
        platform:
          $ref: '#/components/schemas/Platform'
          default: web
      additionalProperties: false
      type: object
      title: Variant
...
    Problem:
      additionalProperties: false
      properties:
        type:
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          title: Detail
          type: string
        code:
          title: Code
          type: string
        instance:
          default: null
          title: Instance
          type: string
          nullable: true
        request_id:
          default: null
          title: Request Id
          type: string
          nullable: true
        errors:
          default: null
          title: Errors
          items:
            additionalProperties: true
            type: object
          type: array
          nullable: true
      required:
      - type
      - title
      - status
      - detail
      - code
      title: Problem
      type: object
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: A scoped key created by an ExitRoute administrator.
      in: header
      name: X-API-Key
    BootstrapAdminAuth:
      type: http
      description: Bootstrap-only bearer token. Prefer a scoped admin API key after setup.
      scheme: bearer
  responses:
...
    Unauthorized:
      description: Authentication failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the required scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unprocessable:
      description: The request violates a semantic or safety rule.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: The API-client quota was exceeded.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Seconds until a new quota window.
          schema:
            type: integer
            minimum: 1
    InternalError:
      description: The server could not complete the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...